+++ rotkehlchen/chain/polygon_pos/modules/drips/v1/decoder.py
@@ -1,0 +1,24 @@
+from typing import TYPE_CHECKING
+from rotkehlchen.chain.evm.decoding.drips.v1.decoder import Dripsv1CommonDecoder
+from rotkehlchen.chain.evm.types import string_to_evm_address
+
+if TYPE_CHECKING:
+    from rotkehlchen.chain.evm.decoding.base import BaseDecoderTools
+    from rotkehlchen.chain.evm.node_inquirer import EvmNodeInquirer
+    from rotkehlchen.user_messages import MessagesAggregator
+
+
+class Dripsv1Decoder(Dripsv1CommonDecoder):
+
+    def __init__(
+            self,
+            evm_inquirer: 'EvmNodeInquirer',
+            base_tools: 'BaseDecoderTools',
+            msg_aggregator: 'MessagesAggregator',
+    ) -> None:
+        super().__init__(
+            evm_inquirer=evm_inquirer,
+            base_tools=base_tools,
+            msg_aggregator=msg_aggregator,
+            drips_hub=string_to_evm_address('0x8715FA4d3831301Ae46EbF19bE37F7a2CA9655D3'),
+        )